Create basic video processing service setup
Initialized video processing service directory, using typescript and
express as the basis of the app. Created the package.json file, added
dependencies and dev dependencies to allow for seamless mixing of node,
TS, and express. Made an index.ts file to serve as the basis for this
portion of the app. Also made a tsconfig JSON file to process the TS
code.